Move antibody transformation to its own file, fix compile errors from merge conflicts
